+++ CHANGES.md
@@ -223,6 +223,14 @@ To be released.
     (``tzinfo``) awareness check for `datetime`, and basic format check for
     `uri`.
 
+ - Generated service methods now have `__nirum_argument_serializers__`
+   attribute, a mapping object that keys are a string of method's parameter
+   facial name and values are its serializer.
+
+   A serializer function takes an argument value and returns its corresponding
+   value which can be passed to `json.dump()`/`json.dumps()`.
+   It can raise a `TypeError` or `ValueError` if an argument is invalid.
+
  -  Removed `__nirum_get_inner_type__()` class methods from generated unboxed
     type classes.